third_party/blink/web_tests/platform/generic/fast/html/set-text-direction-expected.txt

Test that WebKit changes the dir attribute and sends an input event when we change the text direction.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS expected is textarea.dir
PASS sentInputEvent is true
PASS expected is textarea.dir
PASS sentInputEvent is true
PASS expected is input.dir
PASS sentInputEvent is true
PASS expected is input.dir
PASS sentInputEvent is true
PASS successfullyParsed is true

TEST COMPLETE
